Small tweaks again

+++ network.h
@@ -62,7 +62,7 @@ struct client_args {
  * Remainder is text
  * Want the outgoing packet to be 4096 max, aka page size
  */
-#define MAX_INPUT_SIZE 4096 - IV_SIZE - TAG_SIZE - sizeof(uint16_t)
+#define MAX_INPUT_SIZE (uint16_t)(4096 - IV_SIZE - TAG_SIZE - sizeof(uint16_t))
 #define MAX_PACKET_SIZE MAX_INPUT_SIZE + IV_SIZE + TAG_SIZE + sizeof(uint16_t)
 
 extern bool isServer;
